Add List of Available Models to --help
This commit builds off the previous commit to aggregate the outputs from
the 'list_models' method of all current Models. It introduces the
utility method 'get_supported_models', which gathers the supported model
names across all models. This output is then added to the __doc__
variable to be used in the ArgumentParser description.

Issue #224 Add List of Available Models