Add Prefixes to BaseModel
This commit adds the PREFIXES list as a class variable to the BaseModel
class, which is intended to store the potential model name prefixes(if any exist)
for a model name. It also adds the instance method 'list_models', which
first collects the given names in a Model object's MODELS dictionary,
then if any prefixes exist, also collects the model names with the prefix
prepended to the name. Finally, it populates the OpenAIModel and Ollama Models
with the relevant prefixes currently supported.

Issue #224 Add List of Available Models